improve looking game

+++ Functions/Functions_5P.py
@@ -76,7 +76,7 @@ def all_script(driver):
         if passageset:
 
             GetSetActuel(driver)
-            config.newset = config.set_actuel + 1
+            config.newset = int(config.set_actuel) + 1
             config.score_actuel = '0:0'
             gamestart = True
             if config.rattrape_perte == 1:
@@ -190,7 +190,7 @@ def all_script(driver):
             elif str(config.newset) == str(config.set_actuel):  ##SI ON EST SUR LE PROCHAIN SET
                 txtlog = " ON EST SUR LE PROCHAIN SET"
                 passageset = True
-                config.newset = float(config.set_actuel) + 1
+                config.newset = int(config.set_actuel) + 1
                 config.log(txtlog, config.newmatch)
                 DeleteBet(driver)
                 txtlog = 'Wait 30 sec'
